// apps/admin-web/src/api/model/termResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { TermResponseDtoTermType } from './termResponseDtoTermType';

export interface TermResponseDto {
  termType: TermResponseDtoTermType;
  id: string;
  versionId?: string;
  title: string;
  content: string;
  version: string;
  isActive: boolean;
  startDate: string;
  endDate?: string;
  updatedAt: string;
  createdAt: string;
}